pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentracer"
version = "0.1.0"
description = "Lightweight AI cost observability. Know exactly what your AI features cost."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Agentracer", email = "hello@agentracer.dev" }
]
keywords = ["ai", "observability", "llm", "cost-tracking", "openai", "anthropic", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.24.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
gemini = ["google-generativeai>=0.3.0"]
all = ["openai>=1.0.0", "anthropic>=0.18.0", "google-generativeai>=0.3.0"]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[project.urls]
Homepage = "https://agentracer.dev"
Repository = "https://github.com/aqib-ilyas/agentracer-python"
Documentation = "https://agentracer.dev/docs"
